Constructs a legend of paired hamlets with two fills

hypo_legend_pair_split combines several paired hamlet legend elements

hypo_legend_pair_split(left, right, color_map_left, color_map_right,
  circle_color = NA, circle_lwd = 0.5, plot_names = FALSE,
  plot_name_size = 3, font_family = "sans", ncol = 1, plot = TRUE)

Arguments

left
string vector (manatory), can only contain available species

right
string vector (manatory), can only contain available species

color_map_left
color vector (a string, optional), the color map (must match the species vector in length)

color_map_right
color vector (a string, optional), the color map (must match the species vector in length)

circle_color
color skalar (a string, optional), the color of the background circle outline

circle_lwd
numeric skalar (optional), the width of the background circle outlines

plot_names
logical skalar (optional), should the species label be added?

plot_name_size
numeric skalar (optional), the species label size

font_family
string skalar (optional), the species label font family

plot
logical scalar (optional), toggle the output to be either a plot or a list of plots

Details

The function hypo_legend_pair constructs a paired hamlet legend from two vectors of hamlet species and a choosen color map (matching in length).

Hamlet species labels can optionally be included.

Available species: "aberrans", "atlahua", "castroaguirrei", "chlorurus", "ecosur", "floridae", "gemma", "gumigutta", "guttavarius", "indigo", "liberte", "maculiferus", "maya", "nigricans", "providencianus", "puella", "randallorum", "tan", "unicolor"

See also

Examples

clr <- viridis::viridis(4)

left <- c('unicolor', 'liberte', 'maya', 'castroaguirrei')
right <- c('guttavarius', 'gumigutta', 'atlahua', 'randallorum')

hypo_legend_pair(left= left, right = right, color_map = clr,
  circle_color = 'black', plot_names = TRUE)

#> Error in .f(left = .l[[1L]][[i]], right = .l[[2L]][[i]], circle_fill = .l[[3L]][[i]],     ...): right %in% hypo_img$spec is not TRUE